back/requirements.txt
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
google-api-core==2.24.2
google-auth==2.40.0
google-cloud-vision==3.10.1
googleapis-common-protos==1.70.0
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
idna==3.10
numpy==2.2.5
opencv-python==4.11.0.86
proto-plus==1.26.1
protobuf==6.31.0rc2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyserial==3.5
requests==2.32.3
rsa==4.9.1
urllib3==2.4.0
websockets==15.0.1
